refactor(ai): the no-stored-money gate gets its own file
The fitness test landed in ratestore_integration_test.go, which was already
past the 500-line advisory before this branch touched it (545) and would
have left at 603. It also does not belong there: that file proves as-of-date
rate resolution and the pricing arithmetic, while this one asserts the
storage SHAPE those reads run against.

Pure move — ratestore_integration_test.go returns byte-for-byte to its
state on main.
